Move hanlder interface to agent directory
