+++ src/lib/modules/basic-messages/handlers/BasicMessageHandler.ts
@@ -1,4 +1,4 @@
-import { Handler, HandlerInboundMessage } from '../../../handlers/Handler';
+import { Handler, HandlerInboundMessage } from '../../../agent/Handler';
 import { BasicMessageService } from '../services/BasicMessageService';
 import { BasicMessage } from '../messages';
