+++ src/lib/modules/credentials/handlers/CredentialAckHandler.ts
@@ -1,4 +1,4 @@
-import { Handler, HandlerInboundMessage } from '../../../handlers/Handler';
+import { Handler, HandlerInboundMessage } from '../../../agent/Handler';
 import { CredentialService } from '../services';
 import { CredentialAckMessage } from '../messages';
