+++ src/lib/modules/credentials/handlers/IssueCredentialHandler.ts
@@ -1,4 +1,4 @@
-import { Handler, HandlerInboundMessage } from '../../../handlers/Handler';
+import { Handler, HandlerInboundMessage } from '../../../agent/Handler';
 import { CredentialService } from '../services';
 import { IssueCredentialMessage } from '../messages';
