+++ src/lib/modules/credentials/handlers/OfferCredentialHandler.ts
@@ -1,4 +1,4 @@
-import { Handler, HandlerInboundMessage } from '../../../handlers/Handler';
+import { Handler, HandlerInboundMessage } from '../../../agent/Handler';
 import { CredentialService } from '../services';
 import { OfferCredentialMessage } from '../messages';
