+++ src/lib/modules/proofs/handlers/ProposePresentationHandler.ts
@@ -1,4 +1,4 @@
-import { Handler, HandlerInboundMessage } from '../../../handlers/Handler';
+import { Handler, HandlerInboundMessage } from '../../../agent/Handler';
 import { ProposePresentationMessage } from '../messages';
 import { ProofService } from '../services';
